requirements.txt
weaviate-client>=4.9,<5
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-dotenv>=1.0,<2
httpx>=0.27,<1
pydantic>=2,<3
pytest>=8,<9
pytest-asyncio>=0.23,<1
arxiv>=2.4,<3
apscheduler>=3.10,<4
arize-phoenix-otel>=0.15,<1
openinference-semantic-conventions>=0.1,<1
opentelemetry-api>=1.20,<2
opentelemetry-sdk>=1.20,<2
opentelemetry-exporter-otlp-proto-http>=1.20,<2
